+++ inc/posix/private/fd_set.h
@@ -2,9 +2,8 @@
 #define _FD_SET_H
 
 #define FD_SETSIZE 1024
-// TODO: fixme
 typedef struct {
-  unsigned fds_bits[FD_SETSIZE / (sizeof(unsigned) * 8)];
+  unsigned long fds_bits[FD_SETSIZE / (sizeof(unsigned long) * 8)];
 } fd_set;
 
 #endif /* _FD_SET_H */
