+++ inc/posix/private/ino_t.h
@@ -1,6 +1,6 @@
 #ifndef _PRIVATE_INO_T_H
 #define _PRIVATE_INO_T_H
 
-typedef int ino_t; // TODO: look into this type more
+typedef unsigned long long ino_t;
 
 #endif /* _PRIVATE_INO_T_H */
