+++ inc/posix/private/off_t.h
@@ -1,6 +1,6 @@
 #ifndef _PRIVATE_OFF_T_H
 #define _PRIVATE_OFF_T_H
 
-typedef int off_t; // TODO: look into this type more
+typedef long long off_t;
 
 #endif /* _PRIVATE_OFF_T_H */
