+++ inc/posix/private/sa_family_t.h
@@ -1,7 +1,6 @@
 #ifndef _PRIVATE_SA_FAMILY_T_H
 #define _PRIVATE_SA_FAMILY_T_H
 
-/* todo: define me better */
-typedef int sa_family_t;
+typedef unsigned short sa_family_t;
 
 #endif /* _PRIVATE_SA_FAMILY_T_H */
